Stop bots from using splash damage aiming when target in water or for plasma gun
Bots did this with every weapon which is doing splash damage, regardless of the dmg radius, which is ESSENTIAL to give someone the benefits of splash damage (e.g.: Plasmagun dmg_radius 20, is too weak).
Bots did splash aiming always, even if the enemy is in water where splash damage is worthless, because of friction, moreover the enemy can be a few units away from ground, which reduces dmg done by radius.