Allow server to create cvars on client again
If connecting to server with different fs_game, might not of registered the cvars in cgame yet
(so create it so cgame can get the value).

Don't mark server created cvars as read-only (they'll be stuck that way), but still don't let user change them.
Change order of refusing to set cvar so we tell user cvar is server created before read-only etc.