If surface has no lightmap, change lightmap stages to vertex light
If shader has a lightmap stage and surface is vertex lit,
convert stage to vertex light. Fixes a puddle in ET's radar map.

Commit based on ET. I added the isLightmap = qfalse to fix issue
in GL2. It doesn't seem to be an issue in GL1 so added it there too.