Remove ET patch bevel axial plane optimization
It causes false-positives resulting in some axial patch planes not being added to collision detection.

In the Q3 add-on map ct3tourney2 there are two cylinders (made out of patches) that could be walked into (along an axis) due to this change.

I tested the map in ET:Legacy and had the same issue with the cylinders.