spearmint: Fix terminology refering to client but is actually a player
Spearmint separated clients and players because of splitscreen.
Some of the terminology has been changed to reflect this.
A lot of things I had to look at how they were used to figure out what it
was refering too. Hopefully it's less confusing now.

Also fix usage of Com_IsVoipTarget in sv_client.c (because it checked client num,
but should of been player num).