spearmint: Fix systeminfo after fs_game not being set on client
When fs_game was found the variable s was set to value of sv_gameTitle.
This caused a cvar with name from game title to be created and stopped
parsing the system info string. sv_pure and sv_cheats are after fs_game,
so they were not being set. Oops.