spearmint: Improve handling of missing default.cfg
When connecting to a server if default.cfg is missing I made
Spearmint fallback to last valid game and proceed to download mode
which causes a filesystem restart and everythings good. Well, that is,
if the client actually downloaded a pk3 with default.cfg in it.

This commit fixes handling cases where downloads are disabled or
default.cfg is not in the missing pk3s. By fix I mean it gives a sane
error now. This is easy to cause by running a dedicated pure Team Arena
server without putting the required mint-game.settings in a pk3 and
connecting using a client.

Also fixed trying to download files that exist but are not in search path,
which happens in fs_game fallback download mode.