spearmint: Always restart filesystem at connect
If a pk3 is downloaded or fs_game is changed the file system is restarted.
Though if that is not the case, the file system still needs to be restarted
at least once to reload mint-game.settings.

This causes Team Arena single player with sv_pure 1 to require mint-game.settings
to be in a pk3. This is the intended behavior.