spearmint: Add solid EF2 terrain based on Xreal trisoup collide
EF2 terrain is a solid triangle mesh. Xreal has solid trisoup code, so lets use it.
Added the code in cm_patch.c instead of duplicating a lot of code in a new cm_trisoup.c.
Made changes to the Xreal code to make it fit better with the existing code.