spearmint: Fix pure listen server crashing at map_restart
The file system was restarted when client connects so that mint-game.settings
is reloaded. However, when running a server the game VM is already loaded
and saves it's search path to use at map_restart. Restarting the file system
frees and reallocates all the search paths, so the game VM's search path
pointer might not be longer valid anymore. This seems to vary based on
whether mint-game.settings causes pak reordered and/or where game VM
is in the search path (base vs home path, pk3 vs unzipped).

So only restart file system at client connect if not running a server.

Thanks Tobias for steps to reproduce the issue.