spearmint: Reload CGame when demo plays completely
When a demo is playing and player presses a key, disconnect command is run
and cgame is reloaded. However if the entire demo is played, cgame is not
reloaded. CGame doesn't handle demo to main menu transition (not everything
gets cleaned up). It seems best to only have a single case for demo ending;
so let's always reload CGame.