spearmint: Fix shader sort/fog when lightmap stage is first
The auto 'sort' value and fog color adjustment assumed that if there is an
opaque stage if will be listed first. This is not always the case as a
blended lightmap may be listed first.

This commit makes fogging work on some RTCW shaders using 'fog on' that
have a 'blended lightmap', 'opaque' stage, and then a 'blended' stage.