spearmint: Add renderer support for dynamically poly sort order
Add renderer support for dynamically drawing things after others of their
sort value. re.AddPolysToScene can specify a 'sortLevel' which if more than
0 will cause the poly to be drawn after all other shaders with the same
shader sort (SS_DECAL, etc). The sortLevel specifies the order to sort the
polys after the other shaders. This will allow CGame to tell the renderer
what order to draw the wall marks in.

The reason this is nessecary is because by default the wall marks will be
sorted based on the order the shaders were registered. This makes it so marks
are always drawn in a specific order (first registered shader always under
others, etc).

This is needed for World of Padman's spray logos.