spearmint: Clean up NumLock modifier state hacks/handling
Setting modifier state on Linux X11 was committed for SDL 2.0.4.
The Windows handling isn't actually a problem, just ideally it would be
handled by SDL in the future. So there is no point in making it version
dependent. Also checking the current state is the proper thing to do
in case it's changed when the engine is minimized or something.