spearmint: Make customSkin disable surfaces for customShader
If refent has a customSkin, check if it disables model surface before
drawing it with customShader. This makes it so that if a player model
has meshes hidden for some skins, they will only have quad damage effect
if the skin has them enabled. Using customSkin 0 will use the customShader
on all surfaces.

This could be done using a customSkin that uses the custom shader on all
surfaces that are not 'no draw', but this is simpler...