spearmint: Allow adding border to TrueType fonts
Add support for drawing black edge / border around TrueType and OpenType
fonts. The code is based on the freetype example2.cpp by Erik Möller,
which was released into the public domain.

It can be enabled using r_fontBorderWidth cvar (example value: 2).
In the future it will be added to the RegisterFont API.