spearmint: Fix collision with entities in AAS_PredictPlayerMovement
Collision with entities in AAS_PredictPlayerMovement always uses the
same plane (normal 1 0 0) which ruins the bots ability to predict
movement after collision with the entity. The bot would continue trying
to predict moving toward the entity and eventually exit the function
(returning qfalse) because it is stuck in a loop.

Add plane to the AAS trace for storing the entity collision plane
and use it for accessing the trace plane at run-time. The aas trace
planenum might speed up comparisons in places(?) so I've left it for now.

This fixes AAS_PredictPlayerMovement failing to predict movement
on the trap door (activated by a button) in q3dm7.

I added work arounds to avoid breaking compatibility with the Game VM API.