spearmint: Make window resizable by default
Unlike ioquake3 no vid_restart is need for dynamic resizing. The only
issue for Spearmint is that TrueType fonts are rendered to the game
resolution and not updated when the window is resized.

I mainly want to be able to toggle maximized / snap to edge, not manual
arbitry resizing... but I don't think there is a way to only disable the
latter.