spearmint: When loading a demo only display filename not OS path
When loading demos outside of the game file system, the full OS
path was being shown. But the OS path can easily be too long to
fit on the screen. So only show the filename, which also matches
the behavior of loading demos in the game file system.

This only affects the server name given to CGame. The full path can
still be accessed via trap_GetDemoName() in CGame. CGame controls
the text that is draw so it could revert this behavior, display
both, etc.