spearmint: Add cvar to disable sky when using global linear fog
In RTCW the sky is not drawn if there is global linear fog. In ET the sky
is drawn if the shader has stages or sky box. RTCW shaders have broken
stages in sky shaders that should not be drawn. So add a cvar to work
around RTCW's shaders. This keeps the behavior of ET BSP fog and RTCW fog
vars in Spearmint the same.

This change is very noticeable in RTCW's norway map.