spearmint: CGame API 0.35
Add far clip to refdef_t and GetGlobalFog and GetViewFog.

This allows disabling the farclip for specific types of linear fog or
enabling it without linear fog. Now the far clip for the portal skybox
refdef can be disabled to match RTCW behavior.

View fog no longer has a far clip. This makes it more useable now that
it doesn't cause the sky to be visible. Instead the geometry is just
completely fogged beyond the depth to opaque.

The reason water fog still uses far clip is that the RTCW maps (mp_beach
at least) have sky visible underwater (literally a hole in the wall)
and need the clip plane to appear the same as vanilla RTCW.

New maps could use viewfogvars instead of waterfogvars. The benefit of
not using a clip plane is that water can be transparent and the sky is
still visible.