spearmint: Fix cvar default override causing error if same name as VM cmd
"cvarDefault model doom" in mint-game.settings causes model cvar to be
marked as engine created but VMs can only add commands with the same name
as cvars they create (engine drops to menu with error message).

Make cvarDefault in mint-game.settings create cvars that are marked as
user created so that VM takes ownership when it registers the cvar
(allowing command override).

Also allow VM to create commands with the same name as user created cvars
that are not registed by the VM. "spearmint +set say error" >.> lol