spearmint: Add cvar to colorize 2D rgbGen identity for q3test / FAKK
Bitmap font shaders use rgbGen identify. UI graphics use a mix of
rgbGen identity and not specifying -- yet they are colorized by
the older version of the Quake 3 engine. Add a cvar to enable
this behavior.

Also done in FAKK and some games based on it such as McGee's Alice
and Medal of Honor.