spearmint: Don't colorize missing lightmaps by default
q3map2 vertex-approximated lightmap support for Enemy Territory maps
also causes shaders with lightmap stage and no lightmap (that don't
specify surfaceParm nolightmap or that also have surfaceParm
pointlight) to use world vertex lighting which changes the appearence
of some Q3 maps. Such as the Jesus misc_model on q3dm14 and grass
misc_models on tvy-bench.

Some might look okay, but others do not. And really, the shaders
could of used this behavior in the first place if they wanted it.
So it's best to keep the original Q3 behavior by default and add
a new cvar so that ET can use it's not explicitly enabled by the
shader colorizing of lightmap stages.