spearmint: Remove cheat protection from r_lodCurveError
Increasing the value improves visual quaility at a distance. I have
not done much testing but it doesn't seem to give players an
advantage in the game like a cheat.

r_subdivisions 10000 (which is not a cheat) can already reduce the
curve quility at all distances.

It's not a cheat in Enemy Territory, but I don't know when it became
a cheat in Quake 3. Online guides for Q3 suggest changing the
r_lodCurveError cvar which seems to mean r_lodCurveError was not a
cheat originally, though those guides are usually sketchy so who knows.