spearmint: Prevent client from trying to download commericial paks
Make client check built in list of commericial paks and not request
them for download. This fixes only being able to join non-pure server
with auto download disabled on the client when commericial pak is
modified or unzipped.

Don't try to download any paks if missing commericial pak to prevent
downloading addons for a game that cannot be played.