spearmint: Change SDL lib for macOS ppc back to SDL 2.0.3
When SDL was updated to 2.0.4 the macOS ppc lib was downgraded from
modified SDL 2.0.3 to SDL 2.0.1 as it is the last SDL2 version that
officially supports powerpc.

Spearmint requires SDL 2.0.2 for SDL_GameControllerAddMappingsFromFile()
so restore macOS ppc dylib to previously used SDL 2.0.3. macOS SDL lib
for i386 and x86_64 are still SDL 2.0.4.