spearmint: Fix reading/writing demo header
The demo header is aligned to a 4 byte boundary by the compiler. The
header is not fully initialized when it is written, resulting in an
uninitialized byte between magic and headerSize.

I increased the header magic size in the following commit not realizing
the byte was random data.

    commit b2bae95fc00990f5c8b22984ea508401ec5b876f
    Author: Zack Middleton <zack@cloemail.com>
    Date:   Mon Sep 25 22:53:02 2017 -0500
    Subject: spearmint: demoHeader_t comment changes

This broke playing some demos (ones created by non-debug builds?) so
use a padding byte instead of extending the magic size. Also, actually
clear the header so demos don't contain random data.