spearmint: Fix crash in shaderlist command
If a shader exists but there is a "stage with no image", it's stage
pointer and all following stages in the permanent shader are NULL.
It will be marked as defaultShader, so it won't ever be drawn or
anything. Fix my changes to shaderlist that assumed numFoggedPasses
is number of valid stages.