spearmint: Add workaround for client error when using HTTP download
When client connects to server with sv_dlUrl set and tries to download
a pk3, the client errors: ERROR: cgame needs to call trap_SetNetFields.

This can be worked around by adding 8 to sv_allowDownload. That's difficult
to hard code on the server side, so force on the client side.

The issue is that the client sends "disconnect" client command which
causes the server to send entity baseline and snapshot so client can
process the "disconnect" server command. However due to entity and
player data being controlled by CGame, baseline and snapshot can't be
parsed before the correct CGame is loaded (which might not be downloaded
yet).

Reported by mecwerks.