spearmint: Add support for S3Quake3 BSP and DDS files
S3Quake3 is a version of Q3Test that was included with Diamond Viper II
Z200 graphics card. It is a demo for S3 texture compression.

The DDS files have a quirk of incorrect mipmap count specified in header
(all mipmaps exist in the file). The BSP format seems to be the same as
Q3Test 1.06 to 1.08 but with version -46.