spearmint: WIP: Add support for Heavy Metal: FAKK2's skelatal models
The format is also used by American McGee's Alice. Star Trek: Elite
Force 2 uses the same file extensions but it's a slight different
format -- it's not supported by this commit.

This commit draws the models but there is a few issues for full
support.

This only adds it to opengl1 renderer as it's not deemed complete,
merging into opengl2 will not be very difficult. It's missing
support for run-time level-of-detail change, trap_R_LerpTag,
dynamic bones (no CGame API), and legs rotation (I started trying
to reuse WolfET torso rotate API but it's unfinished).
There may be other issues I don't remember right now.

This is partly derived from the WolfET code for MDM+MDX (GPLv3+ with
additional terms) as it was convenient to reuse the already split
mesh and animation files. (FAKK and WolfET models are both based on
MD4 format but collapse map rendering was only in RTCW/WolfET.)